requirements.txt
Django==5.0.3
pillow==10.2.0
simple-salesforce==1.12.5
python-dotenv==1.0.1
pytest-django
django-crispy-forms
crispy-bootstrap4
coverage
django-recaptcha
djangorestframework